增加二级评论功能；增加UITextView+Placeholder扩展；

+++ Weilefood/Helper/ServerHelper/WLServerHelper+Comment.m
@@ -11,7 +11,7 @@
 
 @implementation WLServerHelper (Comment)
 
-- (void)comment_addWithType:(WLCommentType)type refId:(NSUInteger)refId content:(NSString *)content parentId:(NSUInteger)parentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback {
+- (void)comment_addWithType:(WLCommentType)type refId:(long long)refId content:(NSString *)content parentId:(long long)parentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback {
     NSString *apiUrl = [self getApiUrlWithPaths:@[@"comment", @"add"]];
     NSDictionary *parameters = @{@"comment": @{@"Type": @(type),
                                                @"RefId": @(refId),
@@ -21,18 +21,18 @@
     [self httpPOST:apiUrl parameters:parameters callback:callback];
 }
 
-- (void)comment_deleteWithCommentId:(NSUInteger)commentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback {
+- (void)comment_deleteWithCommentId:(long long)commentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback {
     NSString *apiUrl = [self getApiUrlWithPaths:@[@"comment", @"delete"]];
     NSDictionary *parameters = @{@"commentid": @(commentId)};
     [self httpPOST:apiUrl parameters:parameters callback:callback];
 }
 
-- (void)comment_getListWithType:(WLCommentType)type refId:(NSUInteger)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback {
+- (void)comment_getListWithType:(WLCommentType)type refId:(long long)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback {
     NSString *apiUrl = [self getApiUrlWithPaths:@[@"comment", @"list", @(type), @(refId), @(1), @(pageSize), @([maxDate millisecondIntervalSince1970])]];
     [self httpGET:apiUrl parameters:nil resultItemsClass:[WLCommentModel class] callback:callback];
 }
 
-- (void)comment_getMyListWithType:(WLCommentType)type refId:(NSUInteger)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback {
+- (void)comment_getMyListWithType:(WLCommentType)type refId:(long long)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback {
     NSString *apiUrl = [self getApiUrlWithPaths:@[@"comment", @"mylist", @(type), @(refId), @(1), @(pageSize), @([maxDate millisecondIntervalSince1970])]];
     [self httpGET:apiUrl parameters:nil resultItemsClass:[WLCommentModel class] callback:callback];
 }
